update: 更新tools的config配置
